AIX: get rid of common.c, move its functions where they're used
