+++ MANIFEST.in
@@ -26,7 +26,6 @@ include psutil/_psutil_osx.c
 include psutil/_psutil_sunos.c
 include psutil/_psutil_windows.c
 include psutil/_pswindows.py
-include psutil/arch/aix/common.c
 include psutil/arch/aix/common.h
 include psutil/arch/aix/cpu.c
 include psutil/arch/aix/disk.c
